Fix bug in creating samp_pos_to_ref_pos dict
Suppose a sample has a position X in which there is both
a deletion and a SNP (e.g., 'TG' in the reference being 'T'
in this sample can be viewed that way, where the T is deleted
and G -> T). Both will be labeled, in the vphaser output, as
being at position X. But, at the SNP, X will map to position
Y in the reference and, at the deletion, X will map to position
Y-1. This then fails the assertion on
sam_pos_to_ref_pos[row['s_chrom']][samp_pos] because ref_site
changes (it is both Y and Y-1).

This commit changes the position X in the sample, when working
with samp_pos_to_ref_pos, to keep with the existing approach of
subtracting 1 from the position where the allele is a deletion.
So, at the deletion, X becomes X-1 and, at the SNP, X stays X.
This way the assertion does not fail. In the .with-ref-pos.txt.gz
file, the position in the sample will still show X; in the
reference column, the position at the SNP will be Y and in the
deletion it will be Y-1.