Avoid exception when two sequences do not overlap
CoordMapper does not work when there are two sequences that do not overlap.
For example, consider an alignment of length 7 in which there are many
sequences, two of which are:
  --at----
  -----cg-
In this case, CoordMapper2Seqs cannot be instantiated on these
sequences; it raises Exception('CoordMapper2Seqs: no aligned bases.')
because there are no 'real' bases that overlap in the alignment.

This commit allows the object to be created by moving the check from
__init__() to __call__(). Since the sequences do not overlap,
__call__() should not be invoked and, therefore, an exception will
not be raised.