Store vphaser's raw output in tmp directory
This allows re-running the isnvs_per_sample rule without necessarily
having to re-run vphaser (e.g., if only viral-ngs's wrapper around
vphaser's output has changed). It will read the raw output if it
already exists in the directory; otherwise, it will run vphaser and
store the raw output in the tmp directory.

Note that this will only work with genomes that have a single segment.